Fix shell dying
Sometimes the shell dies.
In this cases the next command will not be executed
and the last response from bash will be read from the process' stdout:
"exit".
Since the prompt does not return the thread hangs indefinitely.
This fix polls the shell process before each command and restarts the shell in case the process died.