Do not automatically start CMD, but allow to run it after init
Immediately starting the CMD field is nice for simple, standalone
containers. However, if the containers rely on a topology described by
containernet, their starting commands are likely going to fail because
the initialization of the network did not occur at the time at which the
command is executed.

Thus, giving the user the option to launch the processes at a
well-defined state of the network allows a lot more flexibility.

This change should also remove the need of workarounds like the
`VIM_EMU_CMD` handling in vim-emu.