Don't run containers as privileged
Modified node.py to prevent containers from running as privileged.
This allows namespace isolation to be used. Also hardcoded the net_admin
capability as it is required to run the mininet network; i.e., if a user
attempts to add capabilities, their new capabilities list will always
have net_admin in it.